refactor position const variables and Regexp
