--- src/configs/const.js
@@ -1,49 +1,0 @@
-export const AUTH_ERRORS = {
-  ACCESS_TOKEN_IS_NOT_VALID: 'USER_NOT_AUTHORIZED',
-  REFRESH_TOKEN_IS_NOT_VALID: 'REFRESH_TOKEN_IS_NOT_VALID'
-};
-export const COURIER = 'COURIER';
-export const RESET = 'reset';
-
-export const CURRENCY = {
-  UAH: 'UAH',
-  USD: 'USD'
-};
-export const IMG_ALT = {
-  REGISTER_IMG_INFO: 'register info'
-};
-
-export const languageName = 'language';
-export const limitHistoryOrders = 10;
-
-export const TEXT_FIELD_VARIANT = {
-  OUTLINED: 'outlined',
-  ROUNTED: 'rounded'
-};
-export const TEXT_FIELD_SIZE = {
-  SMALL: 'small'
-};
-export const MATERIAL_UI_COLOR = {
-  INHERIT: 'inherit',
-  PRIMARY: 'primary',
-  ACTION: 'action'
-};
-export const SLIDER_HOME_PAGE = {
-  ARROW: 'arrow',
-  SLIDER: 'slider'
-};
-
-export const statusColors = [
-  { label: 'CREATED', color: '#9e9e9e' },
-  { label: 'CONFIRMED', color: '#4791db' },
-  { label: 'PRODUCED', color: '#81c784' },
-  { label: 'CANCELLED', color: '#e33371' },
-  { label: 'REFUNDED', color: '#ffb74d' },
-  { label: 'SENT', color: '#b39ddb' },
-  { label: 'DELIVERED', color: '#546e7a' }
-];
-
-export const TYPES_CONST = {
-  NUMBER: 'number',
-  STRING: 'string'
-};
