+++ src/containers/cart-header/cart-header.js
@@ -6,11 +6,10 @@ import Badge from '@material-ui/core/Badge';
 import IconButton from '@material-ui/core/IconButton';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import ShoppingBasketIcon from '@material-ui/icons/ShoppingBasket';
-import { MATERIAL_UI_COLOR } from '../../configs/const';
+import { cartKey, MATERIAL_UI_COLOR } from '../../configs';
 
 import { useStyles } from './cart-header.styles';
 import { getCart } from '../../redux/cart/cart.actions';
-import { cartKey } from '../../configs';
 import routes from '../../configs/routes';
 
 const { pathToCart } = routes;
