+++ src/containers/checkout/checkout-form/checkout-form.js
@@ -12,7 +12,13 @@ import { useTranslation } from 'react-i18next';
 import Grid from '@material-ui/core/Grid';
 import DeliveryType from '../delivery-type/delivery-type';
 import { useStyles } from './checkout-form.styles';
-import { CY_CODE_ERR, SESSION_STORAGE, deliveryTypes } from '../../../configs';
+import {
+  CY_CODE_ERR,
+  SESSION_STORAGE,
+  deliveryTypes,
+  TEXT_FIELD_SIZE,
+  TEXT_FIELD_VARIANT
+} from '../../../configs';
 import { calcPriceForCart } from '../../../utils/priceCalculating';
 import Delivery from './delivery';
 import routes from '../../../configs/routes';
@@ -32,7 +38,6 @@ import {
 } from '../../../utils/checkout';
 import { getCurrencySign } from '../../../utils/currency';
 import { validationSchema } from '../../../validators/checkout';
-import { TEXT_FIELD_SIZE, TEXT_FIELD_VARIANT } from '../../../configs/const';
 import {
   clearSessionStorage,
   getFromSessionStorage,
