+++ src/containers/checkout/checkout-form/delivery/nova-post/nova-post.js
@@ -7,8 +7,7 @@ import _ from 'lodash';
 import { useQuery } from '@apollo/client';
 import { useStyles } from './nova-post.styles';
 import { POSTOMAT } from '../../../../../utils/checkout';
-import { CY_CODE_ERR } from '../../../../../configs/index';
-import { MATERIAL_UI_COLOR, TEXT_FIELD_VARIANT, RESET } from '../../../../../configs/const';
+import { CY_CODE_ERR, MATERIAL_UI_COLOR, TEXT_FIELD_VARIANT, RESET } from '../../../../../configs';
 import { getNovaPoshtaCities, getNovaPoshtaWarehouses } from './operations/nova-post.queries.js';
 
 const NovaPost = ({ setFieldValue, errors, touched, values }) => {
