+++ src/containers/checkout/checkout-form/delivery/ukrpost-and-courier/ukrpost-and-courier.js
@@ -5,9 +5,14 @@ import { TextField } from '@material-ui/core';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import Autocomplete from '@material-ui/lab/Autocomplete';
 import { useStyles } from './ukrpost-and-courier.styles';
-import { CY_CODE_ERR, deliveryTypes } from '../../../../../configs';
+import {
+  CY_CODE_ERR,
+  deliveryTypes,
+  RESET,
+  MATERIAL_UI_COLOR,
+  TEXT_FIELD_VARIANT
+} from '../../../../../configs';
 import { courierInputLabels, POST_OFFICE_NUMBER } from '../../../../../utils/checkout';
-import { RESET, MATERIAL_UI_COLOR, TEXT_FIELD_VARIANT } from '../../../../../configs/const';
 import {
   getUkrPoshtaRegions,
   getUkrPoshtaDistricts,
