+++ src/containers/orders/order-history/order-history-pagination/order-history-pagination.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Pagination } from '@material-ui/lab';
 
 import { useStyles } from './order-history-pagination.style';
-import { TEXT_FIELD_VARIANT } from '../../../../configs/const';
+import { TEXT_FIELD_VARIANT } from '../../../../configs';
 
 const OrderHistoryPagination = ({ data }) => {
   const styles = useStyles();
