+++ src/pages/product-list-page/count-per-page/count-per-page.js
@@ -3,8 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { Button, ButtonGroup } from '@material-ui/core';
 import { useHistory, useLocation } from 'react-router';
 import { useStyles } from './count-per-page.styles';
-import { URL_QUERIES_NAME, ITEMS_PER_PAGE } from '../../../configs/index';
-import { TEXT_FIELD_VARIANT } from '../../../configs/const';
+import { URL_QUERIES_NAME, ITEMS_PER_PAGE, TEXT_FIELD_VARIANT } from '../../../configs';
 
 const CountPerPage = () => {
   const { t } = useTranslation();
