+++ src/pages/product-list-page/product-sort/product-sort.js
@@ -5,8 +5,7 @@ import { useTranslation } from 'react-i18next';
 
 import { useStyles } from './product-sort.styles';
 import CountPerPage from '../count-per-page';
-import { URL_QUERIES_NAME, SORT_BY_SELECT_OPTIONS } from '../../../configs';
-import { TEXT_FIELD_VARIANT } from '../../../configs/const';
+import { URL_QUERIES_NAME, SORT_BY_SELECT_OPTIONS, TEXT_FIELD_VARIANT } from '../../../configs';
 
 const ProductSort = () => {
   const { t } = useTranslation();
