+++ src/redux/user/tests/user.sagas.spec.js
@@ -64,9 +64,8 @@ import {
 import routes from '../../../configs/routes';
 import { getCartByUserId, mergeCartFromLSWithUserCart } from '../../cart/cart.operations';
 import { resetCart, setCart, setCartLoading, setCartTotalPrice } from '../../cart/cart.actions';
-import { USER_IS_BLOCKED } from '../../../configs';
+import { USER_IS_BLOCKED, AUTH_ERRORS } from '../../../configs';
 import { handleUserIsBlocked } from '../../../utils/user-helpers';
-import { AUTH_ERRORS } from '../../../configs/const';
 import { USER_ERROR } from '../../../translations/user.translations';
 import { clearLocalStorage } from '../../../services/local-storage.service';
