+++ src/redux/user/user.sagas.js
@@ -49,7 +49,8 @@ import {
   REDIRECT_TIMEOUT,
   RETURN_PAGE,
   USER_IS_BLOCKED,
-  USER_TOKENS
+  USER_TOKENS,
+  AUTH_ERRORS
 } from '../../configs';
 import routes from '../../configs/routes';
 import {
@@ -59,7 +60,6 @@ import {
 } from '../../services/local-storage.service';
 import { resetCart, setCart, setCartLoading, setCartTotalPrice } from '../cart/cart.actions';
 import { handleUserIsBlocked } from '../../utils/user-helpers';
-import { AUTH_ERRORS } from '../../configs/const';
 import { USER_ERROR } from '../../translations/user.translations';
 
 const { pathToLogin, pathToProfile } = routes;
